fix: stub chart component and bump leptos-use after chartistry pulled newer web-sys
Adding leptos-chartistry transitively bumped web-sys, which broke
leptos-use 0.17.0 (geolocation signature drift) and the
HtmlElement::set_scroll_top / scroll_top signatures used in
virtual_scroller. Bump leptos-use to 0.18.3 (the version chartistry
itself uses) and adjust the two scroll-top call sites for the f64
signature.

The PriceHistoryChart component is stubbed to a placeholder until
Task 6 implements the new chartistry chart. The group_sales_by_locale
helper added in Task 2 stays and is exercised by its unit tests.

+++ ultros-frontend/ultros-app/src/components/virtual_scroller.rs
@@ -187,7 +187,7 @@ where
                             // easeOutCubic
                             let ease = 1.0 - (1.0 - t) * (1.0 - t) * (1.0 - t);
                             let val = from + (to - from) * ease;
-                            div_clone.set_scroll_top(val.round() as i32);
+                            div_clone.set_scroll_top(val.round());
                             if t < 1.0 {
                                 if let Some(w) = window() {
                                     let _ = w.request_animation_frame(
@@ -210,7 +210,7 @@ where
                         );
                     } else {
                         // fallback without rAF
-                        div.set_scroll_top(desired.round() as i32);
+                        div.set_scroll_top(desired.round());
                     }
                 }
             }
@@ -237,7 +237,7 @@ where
         <div
             on:scroll=move |scroll| {
                 let div = event_target::<HtmlDivElement>(&scroll);
-                last_scroll.set(div.scroll_top());
+                last_scroll.set(div.scroll_top() as i32);
                 if !raf_pending.get_untracked() {
                     raf_pending.set(true);
                     let last_scroll = last_scroll;
